server/requirements.txt
Django==1.9.5
dj-database-url==0.4.1
django-flatblocks==0.9.2
funcy==1.6
gunicorn==19.4.5
ipdb==0.9.3
ipython==4.2.0
psycopg2==2.6.1
wheel==0.26.0
whitenoise==3.0

django-grappelli==2.8.1
factory-boy==2.4.1
django-wysiwyg-redactor==0.4.9
django-admin-sortable==2.0.15
django-picklefield==0.3.2
django-constance==1.1.2
djangorestframework==3.3.3
xlrd==0.9.4
django-rest-swagger==0.3.6
django-cors-headers==0.12
mock==2.0.0
sorl-thumbnail==12.4a1

requests==2.9.1
lxml==3.6.0
Pillow==3.2.0
coverage==4.0.3
